+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2025-02-25: v0.5.0
 ### Fix
-1. Different ex size in px unit from svg and document env
+1. Different ex size in px unit from svg and document env.
 2. Reduce the default base size because it's too large.
 
 ## 2025-02-25: v0.4.9